+++ src/components/navbar.tsx
@@ -1,8 +1,13 @@
 import { Bell, ChevronDown } from "lucide-react";
 import {DropdownMenu, DropdownMenuContent, DropdownMenuTrigger} from "@/components/ui/dropdown-menu";
 import LogoutButton from "@/components/logoutButton";
+import ProfileButton from "@/components/ProfileButton";
+import {useAuthContext} from "@/context/authContext";
 
 export default function Navbar() {
+    const {user} = useAuthContext();
+    const userId = user?.userId as number;
+
     return (
         <div className="h-16 flex justify-between items-center">
             <div className="text-text-white text-lg">
@@ -12,15 +17,16 @@ export default function Navbar() {
                 <Bell className="text-text-white mr-4" />
                 <DropdownMenu>
                     <DropdownMenuTrigger>
-                <div className="flex items-center bg-tertiary-black p-2 rounded-lg cursor-pointer">
-                    <div
-                        className="w-8 h-8 rounded-full mr-2"
-                    />
-                    <span className="text-text-white">Anne Honyme</span>
-                    <ChevronDown className="text-text-white ml-2" />
-                </div>
+                        <div className="flex items-center bg-tertiary-black p-2 rounded-lg cursor-pointer">
+                            <div
+                                className="w-8 h-8 rounded-full mr-2"
+                            />
+                            <span className="text-text-white">Anne Honyme</span>
+                            <ChevronDown className="text-text-white ml-2" />
+                        </div>
                     </DropdownMenuTrigger>
                     <DropdownMenuContent>
+                        <ProfileButton userId={userId}/>
                         <LogoutButton/>
                     </DropdownMenuContent>
                 </DropdownMenu>